feat(trash_sync): do not restore assets deleted locally only small fixes

+++ mobile/lib/services/action.service.dart
@@ -5,9 +5,13 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:immich_mobile/constants/enums.dart';
 import 'package:immich_mobile/domain/models/asset/base_asset.model.dart';
+import 'package:immich_mobile/domain/models/store.model.dart';
+import 'package:immich_mobile/entities/store.entity.dart';
+import 'package:immich_mobile/extensions/platform_extensions.dart';
 import 'package:immich_mobile/infrastructure/repositories/local_asset.repository.dart';
 import 'package:immich_mobile/infrastructure/repositories/remote_album.repository.dart';
 import 'package:immich_mobile/infrastructure/repositories/remote_asset.repository.dart';
+import 'package:immich_mobile/infrastructure/repositories/trashed_local_asset.repository.dart';
 import 'package:immich_mobile/providers/infrastructure/album.provider.dart';
 import 'package:immich_mobile/providers/infrastructure/asset.provider.dart';
 import 'package:immich_mobile/repositories/asset_api.repository.dart';
@@ -27,6 +31,7 @@ final actionServiceProvider = Provider<ActionService>(
     ref.watch(localAssetRepository),
     ref.watch(driftAlbumApiRepositoryProvider),
     ref.watch(remoteAlbumRepository),
+    ref.watch(trashedLocalAssetRepository),
     ref.watch(assetMediaRepositoryProvider),
     ref.watch(downloadRepositoryProvider),
   ),
@@ -38,6 +43,7 @@ class ActionService {
   final DriftLocalAssetRepository _localAssetRepository;
   final DriftAlbumApiRepository _albumApiRepository;
   final DriftRemoteAlbumRepository _remoteAlbumRepository;
+  final DriftTrashedLocalAssetRepository _trashedLocalAssetRepository;
   final AssetMediaRepository _assetMediaRepository;
   final DownloadRepository _downloadRepository;
 
@@ -47,6 +53,7 @@ class ActionService {
     this._localAssetRepository,
     this._albumApiRepository,
     this._remoteAlbumRepository,
+    this._trashedLocalAssetRepository,
     this._assetMediaRepository,
     this._downloadRepository,
   );
@@ -112,7 +119,11 @@ class ActionService {
       final deletedIds = await _assetMediaRepository.deleteAll(localIds);
 
       if (deletedIds.isNotEmpty) {
-        await _localAssetRepository.delete(deletedIds);
+        if (CurrentPlatform.isAndroid && Store.get(StoreKey.manageLocalMediaAndroid, false)) {
+          await _trashedLocalAssetRepository.applyTrashedAssets(deletedIds);
+        } else {
+          await _localAssetRepository.delete(deletedIds);
+        }
       }
     }
   }
